Desktop and Mobile UI
Steps:
Based off of the design presented in the project proposal, I first created the desktop UI for the main page of our website. The frame used for this layout was for a Macbook Pro screen (1440 x 900). After this was done, I made the mobile version of the UI. This was done using a frame for an iPhone 8 Plus screen (414 x 736). The mobile version has the same contents as the desktop version but is rearranged to fit the different frame.

Usage/Features:
- Within each parking space shown in the parking garage layout is either a green or red light. A green light signals that that space is available while a red light means there is already a vehicle parked in that space. Each space is also identifiable by a letter and number, for example B3, so that it is easier to keep track of which vehicles are in which spots. These lights should be able to update close to real-time.
- Each floor is clickable so that garage layout on the right will change to show the available/taken spots for the corresponding floor.
- Within the checkout section of the website, the user can choose the amount of time they need to be parked and then click checkout. Upon clicking that, the user will be redirected to a checkout page where they enter their full name, email, and credit/debit card information.

Plan:
- create a UI for the checkout page
- work on actual creation of the website

Tools used:
Figma (https://www.figma.com/)